[update] GameTimeを組み込み

+++ GameTime.hpp
@@ -3,18 +3,18 @@
 
 class GameTime
 {
-    const double target_tpf_;
-    double tpf_;
+    const int target_tpf_;
+    int tpf_;
     std::chrono::system_clock::time_point start_time_;
     std::chrono::system_clock::time_point game_time_;
 
     GameTime(const GameTime& other) = delete;
     const GameTime& operator=(const GameTime& other) = delete;
 public:
-    explicit GameTime(const double& target_tpf);
+    explicit GameTime(const int& target_tpf);
     ~GameTime();
     void update();
-    double getTPF();
+    int getTPF();
     void onLoopEnd();
 
 };
